// lib/types.d.ts
export interface Ministry {
    sys: {
        id: string;
        type: string;
        createdAt: string;
        updatedAt: string;
    };
    fields: {
        ministryName: string;
        slug: string;
        description?: string;
    };
}
// ...
export interface ProjectAsset {
    sys: {
        id: string;
        type: string;
    };
    fields: {
        title: string;
        description?: string;
        file: {
            url: string; 
            fileName: string;
            contentType: string; 
            details?: {
                size?: number;
                image?: { width: number; height: number };
            };
        };
    };
}

export interface Project {
    sys: {
        id: string;
        type: string;
        createdAt: string;
        updatedAt: string;
    };
    fields: {
        projectTitle: string;
        projectDescription: string;
        projectImage?: ProjectAsset;
        ministry?: Ministry;
        partners?: ProjectAsset[];
        startDate?: string; 
        proposedEndDate?: string; 
    };
}